+++ dpws/src/test/java/it/org/somda/sdc/dpws/IntegrationTestUtil.java
@@ -5,7 +5,7 @@ import com.google.inject.Injector;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
 import org.somda.sdc.common.guice.DefaultCommonConfigModule;
-import org.somda.sdc.common.guice.DefaultHelperModule;
+import org.somda.sdc.common.guice.DefaultCommonModule;
 import org.somda.sdc.dpws.guice.DefaultDpwsConfigModule;
 import org.somda.sdc.dpws.guice.DefaultDpwsModule;
 import test.org.somda.common.CIDetector;
@@ -34,7 +34,7 @@ public class IntegrationTestUtil {
         injector = Guice.createInjector(
                 new DefaultCommonConfigModule(),
                 new DefaultDpwsModule(),
-                new DefaultHelperModule(),
+                new DefaultCommonModule(),
                 new DefaultDpwsConfigModule());
     }